requirements.txt
alpaca-py
pandas
numpy
matplotlib
python-dotenv
plotly
scikit-learn
requests
